requirements.txt
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.4.1
autopep8==1.5.3
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.14.53
botocore==1.17.53
Brlapi==0.6.6
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
command-not-found==0.3
comment-parser==1.2.2
comtypes==1.1.7
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
fpdf==1.7.2
gensim==3.8.3
gTTS==2.1.1
gTTS-token==1.1.3
httplib2==0.9.2
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lightgbm==3.0.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
netifaces==0.10.4
nltk==3.5
notify2==0.3.1
numpy==1.19.1
oauth==1.0.1
olefile==0.45.1
pandas==1.1.2
pdfkit==0.6.1
pexpect==4.2.1
Pillow==7.2.0
protobuf==3.0.0
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.16.2
pycodestyle==2.6.0
pycricbuzz==2.4
pycrypto==2.6.1
pycryptodome==3.9.8
pycups==1.9.73
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.5.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.8.1
python-debian==0.1.32
python-magic==0.4.18
pyttsx3==2.87
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
regex==2020.7.14
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
simplejson==3.13.2
six==1.15.0
sklearn==0.0
smart-open==2.1.1
soupsieve==2.0
SpeechRecognition==3.8.1
system-service==0.3
systemd-python==234
Telethon==1.11.3
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.48.2
twilio==6.36.0
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.12.1
wxdc==0.1.5
wxPython==4.1.0
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.3.3
zope.interface==4.3.2
